src/requirements.txt
streamlit>=1.32.0
google-generativeai>=0.4.1
pandas>=2.2.1
python-dotenv>=1.0.1
requests>=2.31.0